@@ -88,20 +92,21 @@
     "mock-fs": "^3.11.0",
     "nightwatch": "^0.9.5",
     "open": "0.0.5",
-    "phantomjs": "^1.9.18",
     "proxyquire": "^1.7.10",
     "querystringify": "0.0.3",
     "requirejs": "^2.1.20",
     "sinon": "^1.12.1",
-    "string-replace-webpack-plugin": "0.0.3",
+    "string-replace-webpack-plugin": "^0.1.3",
+    "through2": "^2.0.3",
     "uglify-js": "^2.8.10",
     "url-parse": "^1.0.5",
-    "webpack": "^1.12.3",
-    "webpack-stream": "^3.1.0",
+    "webpack": "^3.0.0",
+    "webpack-stream": "^3.2.0",
     "yargs": "^1.3.1"
   },
   "dependencies": {
     "babel-plugin-transform-object-assign": "^6.22.0",
-    "core-js": "^2.4.1"
+    "core-js": "^2.4.1",
+    "gulp-sourcemaps": "^2.6.0"
   }
 }
