@@ -1,43 +1,51 @@
-var gulp = require('gulp');
+'use strict';
+
+var _ = require('lodash');
 var argv = require('yargs').argv;
+var gulp = require('gulp');
 var gutil = require('gulp-util');
 var connect = require('gulp-connect');
-var webpack = require('webpack-stream');
+var path = require('path');
+var webpack = require('webpack');
+var webpackStream = require('webpack-stream');
 var uglify = require('gulp-uglify');
 var clean = require('gulp-clean');
-var karma = require('gulp-karma');
-var mocha = require('gulp-mocha');
+var KarmaServer = require('karma').Server;
+var karmaConfMaker = require('./karma.conf.maker');
 var opens = require('open');
-var webpackConfig = require('./webpack.conf.js');
+var webpackConfig = require('./webpack.conf');
 var helpers = require('./gulpHelpers');
 var del = require('del');
-var gulpJsdoc2md = require('gulp-jsdoc-to-markdown');
+var gulpDocumentation = require('gulp-documentation');
 var concat = require('gulp-concat');
 var header = require('gulp-header');
-var zip = require('gulp-zip');
+var footer = require('gulp-footer');
 var replace = require('gulp-replace');
 var shell = require('gulp-shell');
 var optimizejs = require('gulp-optimize-js');
-const eslint = require('gulp-eslint');
+var eslint = require('gulp-eslint');
+var gulpif = require('gulp-if');
+var sourcemaps = require('gulp-sourcemaps');
+var through = require('through2');
+var fs = require('fs');
+var jsEscape = require('gulp-js-escape');
 
-var CI_MODE = process.env.NODE_ENV === 'ci';
 var prebid = require('./package.json');
 var dateString = 'Updated : ' + (new Date()).toISOString().substring(0, 10);
-var packageNameVersion = prebid.name + '_' + prebid.version;
 var banner = '/* <%= prebid.name %> v<%= prebid.version %>\n' + dateString + ' */\n';
 var analyticsDirectory = '../analytics';
 var port = 9999;
 
 // Tasks
-gulp.task('default', ['clean', 'lint', 'webpack']);
+gulp.task('default', ['webpack']);
 
-gulp.task('serve', ['clean', 'lint', 'devpack', 'webpack', 'watch', 'test']);
+gulp.task('serve', ['lint', 'build-bundle-dev', 'watch', 'test']);
 
-gulp.task('serve-nw', ['clean', 'lint', 'devpack', 'webpack', 'watch', 'e2etest']);
+gulp.task('serve-nw', ['lint', 'watch', 'e2etest']);
 
-gulp.task('run-tests', ['clean', 'lint', 'webpack', 'test', 'mocha']);
+gulp.task('run-tests', ['lint', 'test-coverage']);
 
-gulp.task('build', ['webpack']);
+gulp.task('build', ['build-bundle-prod']);
 
 gulp.task('clean', function () {
   return gulp.src(['build'], {
@@ -46,146 +54,177 @@ gulp.task('clean', function () {
     .pipe(clean());
 });
 
+function gulpBundle(dev) {
+  return bundle(dev).pipe(gulp.dest('build/' + (dev ? 'dev' : 'dist')));
+}
+
+function nodeBundle(modules) {
+  return new Promise((resolve, reject) => {
+    bundle(false, modules)
+      .on('error', (err) => {
+        reject(err);
+      })
+      .pipe(through.obj(function(file, enc, done) {
+        resolve(file.contents.toString(enc));
+        done();
+      }));
+  });
+}
+
+function bundle(dev, moduleArr) {
+  var modules = moduleArr || helpers.getArgModules(),
+      allModules = helpers.getModuleNames(modules);
+
+  if(modules.length === 0) {
+    modules = allModules;
+  } else {
+    var diff = _.difference(modules, allModules);
+    if(diff.length !== 0) {
+      throw new gutil.PluginError({
+        plugin: 'bundle',
+        message: 'invalid modules: ' + diff.join(', ')
+      });
+    }
+  }
+
+  var entries = [helpers.getBuiltPrebidCoreFile(dev)].concat(helpers.getBuiltModules(dev, modules));
+
+  var outputFileName = argv.bundleName ? argv.bundleName : 'prebid.js';
+
+  // change output filename if argument --tag given
+  if (argv.tag && argv.tag.length) {
+    outputFileName = outputFileName.replace(/\.js$/, `.${argv.tag}.js`);
+  }
+
+  gutil.log('Concatenating files:\n', entries);
+  gutil.log('Appending ' + prebid.globalVarName + '.processQueue();');
+  gutil.log('Generating bundle:', outputFileName);
+
+  return gulp.src(
+      entries
+    )
+    .pipe(gulpif(dev, sourcemaps.init({loadMaps: true})))
+    .pipe(concat(outputFileName))
+    .pipe(gulpif(!argv.manualEnable, footer('\n<%= global %>.processQueue();', {
+        global: prebid.globalVarName
+      }
+    )))
+    .pipe(gulpif(dev, sourcemaps.write('.')));
+}
+
+// Workaround for incompatibility between Karma & gulp callbacks.
+// See https://github.com/karma-runner/gulp-karma/issues/18 for some related discussion.
+function newKarmaCallback(done) {
+  return function (exitCode) {
+    if (exitCode) {
+      done(new Error('Karma tests failed with exit code ' + exitCode));
+    } else {
+      done();
+    }
+  }
+}
+
+gulp.task('build-bundle-dev', ['devpack'], gulpBundle.bind(null, true));
+gulp.task('build-bundle-prod', ['webpack'], gulpBundle.bind(null, false));
+gulp.task('bundle', gulpBundle.bind(null, false)); // used for just concatenating pre-built files with no build step
+
+gulp.task('bundle-to-stdout', function() {
+  nodeBundle().then(file => console.log(file));
+});
+
 gulp.task('devpack', ['clean'], function () {
-  webpackConfig.devtool = 'source-map';
+  var cloned = _.cloneDeep(webpackConfig);
+  cloned.devtool = 'source-map';
+  var externalModules = helpers.getArgModules();
+
   const analyticsSources = helpers.getAnalyticsSources(analyticsDirectory);
-  return gulp.src([].concat(analyticsSources, 'src/prebid.js'))
-    .pipe(webpack(webpackConfig))
+  const moduleSources = helpers.getModulePaths(externalModules);
+
+  return gulp.src([].concat(moduleSources, analyticsSources, 'src/prebid.js'))
+    .pipe(helpers.nameModules(externalModules))
+    .pipe(webpackStream(cloned, webpack))
     .pipe(replace('$prebid.version$', prebid.version))
     .pipe(gulp.dest('build/dev'))
     .pipe(connect.reload());
 });
 
 gulp.task('webpack', ['clean'], function () {
+  var cloned = _.cloneDeep(webpackConfig);
 
-  // change output filename if argument --tag given
-  if (argv.tag && argv.tag.length) {
-    webpackConfig.output.filename = 'prebid.' + argv.tag + '.js';
-  }
+  delete cloned.devtool;
 
-  webpackConfig.devtool = null;
+  var externalModules = helpers.getArgModules();
 
   const analyticsSources = helpers.getAnalyticsSources(analyticsDirectory);
-  return gulp.src([].concat(analyticsSources, 'src/prebid.js'))
-    .pipe(webpack(webpackConfig))
+  const moduleSources = helpers.getModulePaths(externalModules);
+
+  return gulp.src([].concat(moduleSources, analyticsSources, 'src/prebid.js'))
+    .pipe(helpers.nameModules(externalModules))
+    .pipe(webpackStream(cloned, webpack))
     .pipe(replace('$prebid.version$', prebid.version))
     .pipe(uglify())
-    .pipe(header(banner, { prebid: prebid }))
+    .pipe(gulpif(file => file.basename === 'prebid-core.js', header(banner, { prebid: prebid })))
     .pipe(optimizejs())
     .pipe(gulp.dest('build/dist'))
     .pipe(connect.reload());
 });
 
-//zip up for release
-gulp.task('zip', ['clean', 'webpack'], function () {
-  return gulp.src(['build/dist/*', 'integrationExamples/gpt/*'])
-    .pipe(zip(packageNameVersion + '.zip'))
-    .pipe(gulp.dest('./'));
-});
+// Run the unit tests.
+//
+// By default, this runs in headless chrome.
+//
+// If --watch is given, the task will re-run unit tests whenever the source code changes
+// If --file "<path-to-test-file>" is given, the task will only run tests in the specified file.
+// If --browserstack is given, it will run the full suite of currently supported browsers.
+// If --browsers is given, browsers can be chosen explicitly. e.g. --browsers=chrome,firefox,ie9
+gulp.task('test', ['clean'], function (done) {
+  var karmaConf = karmaConfMaker(false, argv.browserstack, argv.watch, argv.file);
 
-// Karma Continuous Testing
-// Pass your browsers by using --browsers=chrome,firefox,ie9
-// Run CI by passing --watch
-gulp.task('test', ['clean'], function () {
-  var defaultBrowsers = CI_MODE ? ['PhantomJS'] : ['Chrome'];
-  var browserArgs = helpers.parseBrowserArgs(argv).map(helpers.toCapitalCase);
-
-  if (process.env.TRAVIS) {
-    browserArgs = ['Chrome_travis_ci'];
+  var browserOverride = helpers.parseBrowserArgs(argv).map(helpers.toCapitalCase);
+  if (browserOverride.length > 0) {
+    karmaConf.browsers = browserOverride;
   }
 
-  if (argv.browserstack) {
-    browserArgs = [
-      'bs_ie_13_windows_10',
-      'bs_ie_11_windows_10',
-      'bs_firefox_46_windows_10',
-      'bs_chrome_51_windows_10',
-      'bs_ie_11_windows_8.1',
-      'bs_firefox_46_windows_8.1',
-      'bs_chrome_51_windows_8.1',
-      'bs_ie_10_windows_8',
-      'bs_firefox_46_windows_8',
-      'bs_chrome_51_windows_8',
-      'bs_ie_11_windows_7',
-      'bs_ie_10_windows_7',
-      'bs_ie_9_windows_7',
-      'bs_firefox_46_windows_7',
-      'bs_chrome_51_windows_7',
-      'bs_safari_9.1_mac_elcapitan',
-      'bs_firefox_46_mac_elcapitan',
-      'bs_chrome_51_mac_elcapitan',
-      'bs_safari_8_mac_yosemite',
-      'bs_firefox_46_mac_yosemite',
-      'bs_chrome_51_mac_yosemite',
-      'bs_safari_7.1_mac_mavericks',
-      'bs_firefox_46_mac_mavericks',
-      'bs_chrome_49_mac_mavericks',
-      'bs_ios_7',
-      'bs_ios_8',
-      'bs_ios_9',
-    ];
-  }
-
-  return gulp.src('lookAtKarmaConfJS')
-    .pipe(karma({
-      browsers: (browserArgs.length > 0) ? browserArgs : defaultBrowsers,
-      configFile: 'karma.conf.js',
-      action: (argv.watch) ? 'watch' : 'run'
-    }));
+  new KarmaServer(karmaConf, newKarmaCallback(done)).start();
 });
 
-//
-// Making this task depend on lint is a bit of a hack. The `run-tests` command is the entrypoint for the CI process,
-// and it needs to run all these tasks together. However, the "lint" and "mocha" tasks explode when used in parallel,
-// resulting in some mysterious "ShellJS: internal error TypeError: Cannot read property 'isFile' of undefined"
-// errors.
-//
-// Gulp doesn't support serial dependencies (until gulp 4.0... which is most likely never coming out)... so we have
-// to trick it by declaring 'lint' as a dependency here. See https://github.com/gulpjs/gulp/blob/master/docs/recipes/running-tasks-in-series.md
-//
-gulp.task('mocha', ['webpack', 'lint'], function() {
-    return gulp.src(['test/spec/loaders/**/*.js'], { read: false })
-        .pipe(mocha({
-          reporter: 'spec',
-          globals: {
-            expect: require('chai').expect
-          }
-        }))
-        .on('error', gutil.log);
+// If --file "<path-to-test-file>" is given, the task will only run tests in the specified file.
+gulp.task('test-coverage', ['clean'], function(done) {
+  new KarmaServer(karmaConfMaker(true, false, false, argv.file), newKarmaCallback(done)).start();
 });
 
-// Small task to load coverage reports in the browser
-gulp.task('coverage', function (done) {
+// View the code coverage report in the browser.
+gulp.task('view-coverage', function (done) {
   var coveragePort = 1999;
 
   connect.server({
-    port: 1999,
-    root: 'build/coverage',
+    port: coveragePort,
+    root: 'build/coverage/karma_html',
     livereload: false
   });
-  opens('http://localhost:' + coveragePort + '/coverage/');
+  opens('http://localhost:' + coveragePort);
   done();
 });
 
-gulp.task('coveralls', ['test'], function() { // 2nd arg is a dependency: 'test' must be finished
+gulp.task('coveralls', ['test-coverage'], function() { // 2nd arg is a dependency: 'test' must be finished
   // first send results of istanbul's test coverage to coveralls.io.
   return gulp.src('gulpfile.js', { read: false }) // You have to give it a file, but you don't
   // have to read it.
-    .pipe(shell('cat build/coverage/lcov/lcov.info | node_modules/coveralls/bin/coveralls.js'));
+    .pipe(shell('cat build/coverage/lcov.info | node_modules/coveralls/bin/coveralls.js'));
 });
 
 // Watch Task with Live Reload
 gulp.task('watch', function () {
-
   gulp.watch([
     'src/**/*.js',
+    'modules/**/*.js',
     'test/spec/**/*.js',
     '!test/spec/loaders/**/*.js'
-  ], ['clean', 'lint', 'webpack', 'devpack', 'test']);
+  ], ['lint', 'build-bundle-dev', 'test']);
   gulp.watch([
     'loaders/**/*.js',
     'test/spec/loaders/**/*.js'
-  ], ['lint', 'mocha']);
+  ], ['lint']);
   connect.server({
     https: argv.https,
     port: port,
@@ -195,7 +234,7 @@ gulp.task('watch', function () {
 });
 
 gulp.task('lint', () => {
-  return gulp.src(['src/**/*.js', 'test/**/*.js'])
+  return gulp.src(['src/**/*.js', 'modules/**/*.js', 'test/**/*.js'])
     .pipe(eslint())
     .pipe(eslint.format('stylish'))
     .pipe(eslint.failAfterError());
@@ -207,15 +246,14 @@ gulp.task('clean-docs', function () {
 
 gulp.task('docs', ['clean-docs'], function () {
   return gulp.src('src/prebid.js')
-    .pipe(concat('readme.md'))
-    .pipe(gulpJsdoc2md())
+    .pipe(gulpDocumentation('md'))
     .on('error', function (err) {
-      gutil.log('jsdoc2md failed:', err.message);
+      gutil.log('`gulp-documentation` failed:', err.message);
     })
     .pipe(gulp.dest('docs'));
 });
 
-gulp.task('e2etest', function() {
+gulp.task('e2etest', ['devpack', 'webpack'], function() {
   var cmdQueue = [];
   if(argv.browserstack) {
     var browsers = require('./browsers.json');
@@ -257,5 +295,23 @@ gulp.task('e2etest-report', function() {
   setTimeout(function() {
     opens('http://localhost:' + reportPort + '/' + targetDestinationDir.slice(2) + '/results.html');
   }, 5000);
-
 });
+
+// This task creates postbid.js. Postbid setup is different from prebid.js
+// More info can be found here http://prebid.org/overview/what-is-post-bid.html
+gulp.task('build-postbid', ['escape-postbid-config'], function() {
+  var fileContent = fs.readFileSync('./build/postbid/postbid-config.js', 'utf8');
+
+  return gulp.src('./integrationExamples/postbid/oas/postbid.js')
+    .pipe(replace('\[%%postbid%%\]', fileContent))
+    .pipe(gulp.dest('build/postbid/'));
+});
+
+// Dependant task for building postbid. It escapes postbid-config file.
+gulp.task('escape-postbid-config', function() {
+  gulp.src('./integrationExamples/postbid/oas/postbid-config.js')
+    .pipe(jsEscape())
+    .pipe(gulp.dest('build/postbid/'));
+});
+
+module.exports = nodeBundle;
