@@ -1,8 +1,6 @@
-import { getBidRequest, logError, insertPixel } from './utils';
+import { deepAccess, getBidRequest, logError, triggerPixel } from './utils';
 
-/** INSERT NATIVE ADAPTERS - DO NOT EDIT OR REMOVE */
-const nativeAdapters = [];
-/** END INSERT NATIVE ADAPTERS */
+export const nativeAdapters = [];
 
 export const NATIVE_KEYS = {
   title: 'hb_native_title',
@@ -11,6 +9,7 @@ export const NATIVE_KEYS = {
   image: 'hb_native_image',
   icon: 'hb_native_icon',
   clickUrl: 'hb_native_linkurl',
+  cta: 'hb_native_cta',
 };
 
 export const NATIVE_TARGETING_KEYS = Object.keys(NATIVE_KEYS).map(
@@ -18,12 +17,12 @@ export const NATIVE_TARGETING_KEYS = Object.keys(NATIVE_KEYS).map(
 );
 
 const IMAGE = {
-  image: {required: true},
-  title: {required: true},
-  sponsoredBy: {required: true},
-  clickUrl: {required: true},
-  body: {required: false},
-  icon: {required: false},
+  image: { required: true },
+  title: { required: true },
+  sponsoredBy: { required: true },
+  clickUrl: { required: true },
+  body: { required: false },
+  icon: { required: false },
 };
 
 const SUPPORTED_TYPES = {
@@ -35,7 +34,7 @@ const SUPPORTED_TYPES = {
  * passes them on directly. If they were of type 'type', translate
  * them into the predefined specific asset requests for that type of native ad.
  */
-export default function processNativeAdUnitParams(params) {
+export function processNativeAdUnitParams(params) {
   if (params && params.type && typeIsSupported(params.type)) {
     return SUPPORTED_TYPES[params.type];
   }
@@ -60,7 +59,11 @@ function typeIsSupported(type) {
  * TODO: abstract this and the video helper functions into general
  * adunit validation helper functions
  */
-export const nativeAdUnit = adUnit => adUnit.mediaType === 'native';
+export const nativeAdUnit = adUnit => {
+  const mediaType = adUnit.mediaType === 'native';
+  const mediaTypes = deepAccess(adUnit, 'mediaTypes.native');
+  return mediaType || mediaTypes;
+}
 export const nativeBidder = bid => nativeAdapters.includes(bid.bidder);
 export const hasNonNativeBidder = adUnit =>
   adUnit.bids.filter(bid => !nativeBidder(bid)).length;
@@ -71,28 +74,83 @@ export const hasNonNativeBidder = adUnit =>
  */
 export function nativeBidIsValid(bid) {
   const bidRequest = getBidRequest(bid.adId);
-  if (!bidRequest) { return false; }
+  if (!bidRequest) {
+    return false;
+  }
+
+  // all native bid responses must define a landing page url
+  if (!deepAccess(bid, 'native.clickUrl')) {
+    return false;
+  }
 
   const requestedAssets = bidRequest.nativeParams;
-  if (!requestedAssets) { return true; }
+  if (!requestedAssets) {
+    return true;
+  }
 
   const requiredAssets = Object.keys(requestedAssets).filter(
     key => requestedAssets[key].required
   );
-  const returnedAssets = Object.keys(bid.native);
+  const returnedAssets = Object.keys(bid['native']).filter(
+    key => bid['native'][key]
+  );
 
   return requiredAssets.every(asset => returnedAssets.includes(asset));
 }
 
 /*
- * Native responses may have impression trackers. This retrieves the
- * impression tracker urls for the given ad object and fires them.
+ * Native responses may have associated impression or click trackers.
+ * This retrieves the appropriate tracker urls for the given ad object and
+ * fires them. As a native creatives may be in a cross-origin frame, it may be
+ * necessary to invoke this function via postMessage. secureCreatives is
+ * configured to fire this function when it receives a `message` of 'Prebid Native'
+ * and an `adId` with the value of the `bid.adId`. When a message is posted with
+ * these parameters, impression trackers are fired. To fire click trackers, the
+ * message should contain an `action` set to 'click'.
+ *
+ * // Native creative template example usage
+ * <a href="%%CLICK_URL_UNESC%%%%PATTERN:hb_native_linkurl%%"
+ *    target="_blank"
+ *    onclick="fireTrackers('click')">
+ *    %%PATTERN:hb_native_title%%
+ * </a>
+ *
+ * <script>
+ *   function fireTrackers(action) {
+ *     var message = {message: 'Prebid Native', adId: '%%PATTERN:hb_adid%%'};
+ *     if (action === 'click') {message.action = 'click';} // fires click trackers
+ *     window.parent.postMessage(JSON.stringify(message), '*');
+ *   }
+ *   fireTrackers(); // fires impressions when creative is loaded
+ * </script>
  */
-export function fireNativeImpressions(adObject) {
-  const impressionTrackers = adObject.native &&
-    adObject.native.impressionTrackers;
+export function fireNativeTrackers(message, adObject) {
+  let trackers;
 
-  (impressionTrackers || []).forEach(tracker => {
-    insertPixel(tracker);
+  if (message.action === 'click') {
+    trackers = adObject['native'] && adObject['native'].clickTrackers;
+  } else {
+    trackers = adObject['native'] && adObject['native'].impressionTrackers;
+  }
+
+  (trackers || []).forEach(triggerPixel);
+}
+
+/**
+ * Gets native targeting key-value paris
+ * @param {Object} bid
+ * @return {Object} targeting
+ */
+export function getNativeTargeting(bid) {
+  let keyValues = {};
+
+  Object.keys(bid['native']).forEach(asset => {
+    const key = NATIVE_KEYS[asset];
+    const value = bid['native'][asset];
+    if (key) {
+      keyValues[key] = value;
+    }
   });
+
+  return keyValues;
 }
