@@ -7,6 +7,7 @@ import {
   getTargetingKeysBidLandscape,
   getAdUnits
 } from 'test/fixtures/fixtures';
+import { config as configObj } from 'src/config';
 
 var assert = require('chai').assert;
 var expect = require('chai').expect;
@@ -23,6 +24,10 @@ var events = require('src/events');
 var adserver = require('src/adserver');
 var CONSTANTS = require('src/constants.json');
 
+// These bid adapters are required to be loaded for the following tests to work
+require('modules/appnexusAstBidAdapter');
+require('modules/adequantBidAdapter');
+
 var config = require('test/fixtures/config.json');
 
 $$PREBID_GLOBAL$$ = $$PREBID_GLOBAL$$ || {};
@@ -32,7 +37,7 @@ $$PREBID_GLOBAL$$.adUnits = getAdUnits();
 $$PREBID_GLOBAL$$._adUnitCodes = $$PREBID_GLOBAL$$.adUnits.map(unit => unit.code);
 
 function resetAuction() {
-  $$PREBID_GLOBAL$$._sendAllBids = false;
+  $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: false });
   $$PREBID_GLOBAL$$.clearAuction();
   $$PREBID_GLOBAL$$._bidsRequested = getBidRequests();
   $$PREBID_GLOBAL$$._bidsReceived = getBidResponses();
@@ -135,7 +140,7 @@ window.apntag = {
 describe('Unit: Prebid Module', function () {
   after(function() {
     $$PREBID_GLOBAL$$.adUnits = [];
-  })
+  });
   describe('getAdserverTargetingForAdUnitCodeStr', function () {
     beforeEach(() => {
       resetAuction();
@@ -143,7 +148,7 @@ describe('Unit: Prebid Module', function () {
 
     it('should return targeting info as a string', function () {
       const adUnitCode = config.adUnitCodes[0];
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       var expected = 'foobar=300x250&hb_size=300x250&hb_pb=10.00&hb_adid=233bcbee889d46d&hb_bidder=appnexus&hb_size_triplelift=0x0&hb_pb_triplelift=10.00&hb_adid_triplelift=222bb26f9e8bd&hb_bidder_triplelift=triplelift&hb_size_appnexus=300x250&hb_pb_appnexus=10.00&hb_adid_appnexus=233bcbee889d46d&hb_bidder_appnexus=appnexus&hb_size_pagescience=300x250&hb_pb_pagescience=10.00&hb_adid_pagescience=25bedd4813632d7&hb_bidder_pagescienc=pagescience&hb_size_brightcom=300x250&hb_pb_brightcom=10.00&hb_adid_brightcom=26e0795ab963896&hb_bidder_brightcom=brightcom&hb_size_brealtime=300x250&hb_pb_brealtime=10.00&hb_adid_brealtime=275bd666f5a5a5d&hb_bidder_brealtime=brealtime&hb_size_pubmatic=300x250&hb_pb_pubmatic=10.00&hb_adid_pubmatic=28f4039c636b6a7&hb_bidder_pubmatic=pubmatic&hb_size_rubicon=300x600&hb_pb_rubicon=10.00&hb_adid_rubicon=29019e2ab586a5a&hb_bidder_rubicon=rubicon';
       var result = $$PREBID_GLOBAL$$.getAdserverTargetingForAdUnitCodeStr(adUnitCode);
       assert.equal(expected, result, 'returns expected string of ad targeting info');
@@ -161,7 +166,7 @@ describe('Unit: Prebid Module', function () {
   describe('getAdserverTargetingForAdUnitCode', function () {
     it('should return targeting info as an object', function () {
       const adUnitCode = config.adUnitCodes[0];
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       var result = $$PREBID_GLOBAL$$.getAdserverTargetingForAdUnitCode(adUnitCode);
       const expected = getAdServerTargeting()[adUnitCode];
       assert.deepEqual(result, expected, 'returns expected' +
@@ -179,7 +184,7 @@ describe('Unit: Prebid Module', function () {
     });
 
     it('should return current targeting data for slots', function () {
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       const targeting = $$PREBID_GLOBAL$$.getAdserverTargeting();
       const expected = getAdServerTargeting();
       assert.deepEqual(targeting, expected, 'targeting ok');
@@ -207,7 +212,7 @@ describe('Unit: Prebid Module', function () {
     });
 
     it('should return correct targeting with bid landscape targeting on', () => {
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       var targeting = $$PREBID_GLOBAL$$.getAdserverTargeting();
       var expected = getAdServerTargeting();
       assert.deepEqual(targeting, expected);
@@ -334,6 +339,62 @@ describe('Unit: Prebid Module', function () {
     });
   });
 
+  describe('getAdserverTargeting', function() {
+    const customConfigObject = {
+      'buckets': [
+        { 'precision': 2, 'min': 0, 'max': 5, 'increment': 0.01 },
+        { 'precision': 2, 'min': 5, 'max': 8, 'increment': 0.05},
+        { 'precision': 2, 'min': 8, 'max': 20, 'increment': 0.5 },
+        { 'precision': 2, 'min': 20, 'max': 25, 'increment': 1 }
+      ]
+    };
+    let currentPriceBucket;
+    let bid;
+
+    before(() => {
+      resetAuction();
+      currentPriceBucket = configObj.getConfig('priceGranularity');
+      configObj.setConfig({ priceGranularity: customConfigObject });
+      bid = Object.assign({},
+        bidfactory.createBid(2),
+        getBidResponses()[5]
+      );
+    });
+
+    after(() => {
+      configObj.setConfig({ priceGranularity: currentPriceBucket });
+      resetAuction();
+    })
+
+    beforeEach(() => {
+      $$PREBID_GLOBAL$$._bidsReceived = [];
+    })
+
+    it('should get correct hb_pb when using bid.cpm is between 0 to 5', () => {
+      bid.cpm = 2.1234;
+      bidmanager.addBidResponse(bid.adUnitCode, bid);
+      expect($$PREBID_GLOBAL$$.getAdserverTargeting()['/19968336/header-bid-tag-0'].hb_pb).to.equal('2.12');
+    });
+
+    it('should get correct hb_pb when using bid.cpm is between 5 to 8', () => {
+      bid.cpm = 6.78;
+      bidmanager.addBidResponse(bid.adUnitCode, bid);
+      expect($$PREBID_GLOBAL$$.getAdserverTargeting()['/19968336/header-bid-tag-0'].hb_pb).to.equal('6.75');
+    });
+
+    it('should get correct hb_pb when using bid.cpm is between 8 to 20', () => {
+      bid.cpm = 19.5234;
+      bidmanager.addBidResponse(bid.adUnitCode, bid);
+      expect($$PREBID_GLOBAL$$.getAdserverTargeting()['/19968336/header-bid-tag-0'].hb_pb).to.equal('19.50');
+    });
+
+    it('should get correct hb_pb when using bid.cpm is between 20 to 25', () => {
+      bid.cpm = 21.5234;
+      bidmanager.addBidResponse(bid.adUnitCode, bid);
+      expect($$PREBID_GLOBAL$$.getAdserverTargeting()['/19968336/header-bid-tag-0'].hb_pb).to.equal('21.00');
+    });
+  });
+
   describe('getBidResponses', function () {
     it('should return expected bid responses when not passed an adunitCode', function () {
       var result = $$PREBID_GLOBAL$$.getBidResponses();
@@ -389,7 +450,7 @@ describe('Unit: Prebid Module', function () {
     it('should set targeting when passed a string ad unit code with enableSendAllBids', function () {
       var slots = createSlotArray();
       window.googletag.pubads().setSlots(slots);
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
 
       $$PREBID_GLOBAL$$.setTargetingForGPTAsync('/19968336/header-bid-tag-0');
       expect(slots[0].spySetTargeting.args).to.deep.contain.members([['hb_bidder', 'appnexus'], ['hb_adid_appnexus', '233bcbee889d46d'], ['hb_pb_appnexus', '10.00']]);
@@ -398,7 +459,7 @@ describe('Unit: Prebid Module', function () {
     it('should set targeting when passed an array of ad unit codes with enableSendAllBids', function () {
       var slots = createSlotArray();
       window.googletag.pubads().setSlots(slots);
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
 
       $$PREBID_GLOBAL$$.setTargetingForGPTAsync(['/19968336/header-bid-tag-0']);
       expect(slots[0].spySetTargeting.args).to.deep.contain.members([['hb_bidder', 'appnexus'], ['hb_adid_appnexus', '233bcbee889d46d'], ['hb_pb_appnexus', '10.00']]);
@@ -419,7 +480,7 @@ describe('Unit: Prebid Module', function () {
       var slots = createSlotArray();
       window.googletag.pubads().setSlots(slots);
 
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       $$PREBID_GLOBAL$$.setTargetingForGPTAsync();
 
       var expected = getTargetingKeysBidLandscape();
@@ -718,7 +779,7 @@ describe('Unit: Prebid Module', function () {
       adaptermanager.callBids.restore();
     });
 
-    it('should not callBids if a video adUnit has non-video bidders', () => {
+    it('should only request video bidders on video adunits', () => {
       sinon.spy(adaptermanager, 'callBids');
       const videoAdaptersBackup = adaptermanager.videoAdapters;
       adaptermanager.videoAdapters = ['appnexusAst'];
@@ -732,7 +793,35 @@ describe('Unit: Prebid Module', function () {
       }];
 
       $$PREBID_GLOBAL$$.requestBids({adUnits});
-      sinon.assert.notCalled(adaptermanager.callBids);
+      sinon.assert.calledOnce(adaptermanager.callBids);
+
+      const spyArgs = adaptermanager.callBids.getCall(0);
+      const biddersCalled = spyArgs.args[0].adUnits[0].bids;
+      expect(biddersCalled.length).to.equal(1);
+
+      adaptermanager.callBids.restore();
+      adaptermanager.videoAdapters = videoAdaptersBackup;
+    });
+
+    it('should only request video bidders on video adunits configured with mediaTypes', () => {
+      sinon.spy(adaptermanager, 'callBids');
+      const videoAdaptersBackup = adaptermanager.videoAdapters;
+      adaptermanager.videoAdapters = ['appnexusAst'];
+      const adUnits = [{
+        code: 'adUnit-code',
+        mediaTypes: {video: {context: 'instream'}},
+        bids: [
+          {bidder: 'appnexus', params: {placementId: 'id'}},
+          {bidder: 'appnexusAst', params: {placementId: 'id'}}
+        ]
+      }];
+
+      $$PREBID_GLOBAL$$.requestBids({adUnits});
+      sinon.assert.calledOnce(adaptermanager.callBids);
+
+      const spyArgs = adaptermanager.callBids.getCall(0);
+      const biddersCalled = spyArgs.args[0].adUnits[0].bids;
+      expect(biddersCalled.length).to.equal(1);
 
       adaptermanager.callBids.restore();
       adaptermanager.videoAdapters = videoAdaptersBackup;
@@ -1252,18 +1341,9 @@ describe('Unit: Prebid Module', function () {
       const logErrorSpy = sinon.spy(utils, 'logError');
       const error = 'Prebid Error: no value passed to `setPriceGranularity()`';
 
-      $$PREBID_GLOBAL$$.setPriceGranularity();
+      $$PREBID_GLOBAL$$.setConfig({ priceGranularity: null });
       assert.ok(logErrorSpy.calledWith(error), 'expected error was logged');
       utils.logError.restore();
-    });
-
-    it('should call bidmanager.setPriceGranularity with granularity', () => {
-      const setPriceGranularitySpy = sinon.spy(bidmanager, 'setPriceGranularity');
-      const granularity = 'low';
-
-      $$PREBID_GLOBAL$$.setPriceGranularity(granularity);
-      assert.ok(setPriceGranularitySpy.called, 'called bidmanager.setPriceGranularity');
-      bidmanager.setPriceGranularity.restore();
     });
 
     it('should log error when not passed a valid config object', () => {
@@ -1276,7 +1356,7 @@ describe('Unit: Prebid Module', function () {
           'increment': 0.01,
         },
         {
-            // missing min prop
+          // missing min prop
           'max': 18,
           'increment': 0.05,
           'cap': true
@@ -1284,14 +1364,13 @@ describe('Unit: Prebid Module', function () {
         ]
       };
 
-      $$PREBID_GLOBAL$$.setPriceGranularity(badConfig);
+      $$PREBID_GLOBAL$$.setConfig({ priceGranularity: badConfig });
       assert.ok(logErrorSpy.calledWith(error), 'expected error was logged');
       utils.logError.restore();
     });
 
-    it('should call bidmanager.setCustomPriceBucket with custom config buckets', () => {
-      const setCustomPriceBucket = sinon.spy(bidmanager, 'setCustomPriceBucket');
-      const setPriceGranularitySpy = sinon.spy(bidmanager, 'setPriceGranularity');
+    it('should set customPriceBucket with custom config buckets', () => {
+      let customPriceBucket = configObj.getConfig('customPriceBucket');
       const goodConfig = {
         'buckets': [{
           'min': 0,
@@ -1301,12 +1380,11 @@ describe('Unit: Prebid Module', function () {
         }
         ]
       };
-
-      $$PREBID_GLOBAL$$.setPriceGranularity(goodConfig);
-      assert.ok(setCustomPriceBucket.called, 'called bidmanager.setCustomPriceBucket');
-      bidmanager.setCustomPriceBucket.restore();
-      assert.ok(setPriceGranularitySpy.calledWith('custom'), 'called bidmanager.setPriceGranularity');
-      bidmanager.setPriceGranularity.restore();
+      configObj.setConfig({ priceGranularity: goodConfig });
+      let priceGranularity = configObj.getConfig('priceGranularity');
+      let newCustomPriceBucket = configObj.getConfig('customPriceBucket');
+      expect(goodConfig).to.deep.equal(newCustomPriceBucket);
+      expect(priceGranularity).to.equal(CONSTANTS.GRANULARITY_OPTIONS.CUSTOM);
     });
   });
 
@@ -1577,11 +1655,15 @@ describe('Unit: Prebid Module', function () {
     });
 
     it('should return original adservertag if there are no bids for the given placement code', () => {
-      var options = {
+      // urls.js:parse returns port 443 for IE11, blank for other browsers
+      const ie11port = !!window.MSInputMethodContext && !!document.documentMode ? ':443' : '';
+      const adserverTag = `https://pubads.g.doubleclick.net${ie11port}/gampad/ads?sz=640x480&iu=/19968336/header-bid-tag-0&impl=s&gdfp_req=1&env=vp&output=xml_vast2&unviewed_position_start=1&url=www.test.com`;
+
+      const masterTagUrl = $$PREBID_GLOBAL$$.buildMasterVideoTagFromAdserverTag(adserverTag, {
         'adserver': 'dfp',
         'code': 'one-without-bids'
-      };
-      var masterTagUrl = $$PREBID_GLOBAL$$.buildMasterVideoTagFromAdserverTag(adserverTag, options);
+      });
+
       expect(masterTagUrl).to.equal(adserverTag);
     });
 
@@ -1604,7 +1686,7 @@ describe('Unit: Prebid Module', function () {
     });
   });
 
-  describe('setBidderSequence', () => {
+  describe('bidderSequence', () => {
     it('setting to `random` uses shuffled order of adUnits', () => {
       sinon.spy(utils, 'shuffle');
       const requestObj = {
@@ -1612,7 +1694,7 @@ describe('Unit: Prebid Module', function () {
         timeout: 2000
       };
 
-      $$PREBID_GLOBAL$$.setBidderSequence('random');
+      $$PREBID_GLOBAL$$.setConfig({ bidderSequence: 'random' });
       $$PREBID_GLOBAL$$.requestBids(requestObj);
 
       sinon.assert.calledOnce(utils.shuffle);
@@ -1676,7 +1758,7 @@ describe('Unit: Prebid Module', function () {
 
     it('should not find hb_adid key in lowercase for all bidders', () => {
       const adUnitCode = '/19968336/header-bid-tag-0';
-      $$PREBID_GLOBAL$$.enableSendAllBids();
+      $$PREBID_GLOBAL$$.setConfig({ enableSendAllBids: true });
       $$PREBID_GLOBAL$$.setTargetingForAst();
       const keywords = Object.keys(window.apntag.tags[adUnitCode].keywords).filter(keyword => (keyword.substring(0, 'hb_adid'.length) === 'hb_adid'));
       expect(keywords.length).to.equal(0);
@@ -1737,7 +1819,7 @@ describe('Unit: Prebid Module', function () {
         endpoint: 'https://prebid.adnxs.com/pbs/v1/auction'
       };
 
-      $$PREBID_GLOBAL$$.setS2SConfig(options);
+      $$PREBID_GLOBAL$$.setConfig({ s2sConfig: {options} });
       assert.ok(logErrorSpy.calledOnce, true);
     });
 
@@ -1750,7 +1832,7 @@ describe('Unit: Prebid Module', function () {
         endpoint: 'https://prebid.adnxs.com/pbs/v1/auction'
       };
 
-      $$PREBID_GLOBAL$$.setS2SConfig(options);
+      $$PREBID_GLOBAL$$.setConfig({ s2sConfig: {options} });
       assert.ok(logErrorSpy.calledOnce, true);
     });
   });
