@@ -521,7 +521,173 @@ describe('Utils', function () {
       for (var key in arr) {
         count++;
       }
-      assert.equal(arr.length, count, 'Polyfill test fails')
+      assert.equal(arr.length, count, 'Polyfill test fails');
+    });
+  });
+
+  /**
+   *  tests fail in IE10 because __lookupSetter__ and __lookupGetter__ are
+   *  not supported. See #1656. commenting out until they can be fixed.
+   *
+   *  describe('cookie support', function () {
+   *    // store original cookie getter and setter so we can reset later
+   *    var origCookieSetter = document.__lookupSetter__('cookie');
+   *    var origCookieGetter = document.__lookupGetter__('cookie');
+   *
+   *    // store original cookieEnabled getter and setter so we can reset later
+   *    var origCookieEnabledSetter = window.navigator.__lookupSetter__('cookieEnabled');
+   *    var origCookieEnabledGetter = window.navigator.__lookupGetter__('cookieEnabled');
+   *
+   *    // Replace the document cookie set function with the output of a custom function for testing
+   *    let setCookie = (v) => v;
+   *
+   *    beforeEach(() => {
+   *      // Redefine window.navigator.cookieEnabled such that you can set otherwise "read-only" values
+   *      Object.defineProperty(window.navigator, 'cookieEnabled', (function (_value) {
+   *        return {
+   *          get: function _get() {
+   *            return _value;
+   *          },
+   *          set: function _set(v) {
+   *            _value = v;
+   *          },
+   *          configurable: true
+   *        };
+   *      })(window.navigator.cookieEnabled));
+   *
+   *      // Reset the setCookie cookie function before each test
+   *      setCookie = (v) => v;
+   *      // Redefine the document.cookie object such that you can purposefully have it output nothing as if it is disabled
+   *      Object.defineProperty(window.document, 'cookie', (function (_value) {
+   *        return {
+   *          get: function _get() {
+   *            return _value;
+   *          },
+   *          set: function _set(v) {
+   *            _value = setCookie(v);
+   *          },
+   *          configurable: true
+   *        };
+   *      })(window.navigator.cookieEnabled));
+   *    });
+   *
+   *    afterEach(() => {
+   *      // redefine window.navigator.cookieEnabled to original getter and setter
+   *      Object.defineProperty(window.navigator, 'cookieEnabled', {
+   *        get: origCookieEnabledGetter,
+   *        set: origCookieEnabledSetter,
+   *        configurable: true
+   *      });
+   *      // redefine document.cookie to original getter and setter
+   *      Object.defineProperty(document, 'cookie', {
+   *        get: origCookieGetter,
+   *        set: origCookieSetter,
+   *        configurable: true
+   *      });
+   *    });
+   *
+   *    it('should be detected', function() {
+   *      assert.equal(utils.cookiesAreEnabled(), true, 'Cookies should be enabled by default');
+   *    });
+   *
+   *    it('should be not available', function() {
+   *      setCookie = () => '';
+   *      window.navigator.cookieEnabled = false;
+   *      window.document.cookie = '';
+   *      assert.equal(utils.cookiesAreEnabled(), false, 'Cookies should be disabled');
+   *    });
+   *
+   *    it('should be available', function() {
+   *      window.navigator.cookieEnabled = false;
+   *      window.document.cookie = 'key=value';
+   *      assert.equal(utils.cookiesAreEnabled(), true, 'Cookies should already be set');
+   *      window.navigator.cookieEnabled = false;
+   *      window.document.cookie = '';
+   *      assert.equal(utils.cookiesAreEnabled(), true, 'Cookies should settable');
+   *      setCookie = () => '';
+   *      window.navigator.cookieEnabled = true;
+   *      window.document.cookie = '';
+   *      assert.equal(utils.cookiesAreEnabled(), true, 'Cookies should be on via on window.navigator');
+   *      // Reset the setCookie
+   *      setCookie = (v) => v;
+   *    });
+   *  });
+  **/
+
+  describe('delayExecution', function () {
+    it('should execute the core function after the correct number of calls', function () {
+      const callback = sinon.spy();
+      const delayed = utils.delayExecution(callback, 5);
+      for (let i = 0; i < 4; i++) {
+        delayed();
+      }
+      assert(callback.notCalled);
+      delayed(3);
+      assert(callback.called)
+      assert.equal(callback.firstCall.args[0], 3);
+    });
+  });
+
+  describe('deepAccess', function() {
+    var obj = {
+      1: 2,
+      test: {
+        first: 11
+      }
+    };
+
+    it('should allow deep access of object properties', function() {
+      var value1 = utils.deepAccess(obj, 'test');
+      assert.deepEqual(value1, obj.test);
+
+      var value2 = utils.deepAccess(obj, 'test.first');
+      assert.equal(value2, 11);
+
+      var value3 = utils.deepAccess(obj, 1);
+      assert.equal(value3, 2);
+    });
+
+    it('should allow safe access (returning undefined for missing properties and not throwing exceptions)', function() {
+      var value;
+
+      assert.doesNotThrow(function() {
+        value = utils.deepAccess(obj, 'test.second.third');
+      });
+
+      assert.equal(value, undefined);
+    });
+  });
+
+  describe('createContentToExecuteExtScriptInFriendlyFrame', function () {
+    it('should return empty string if url is not passed', function () {
+      var output = utils.createContentToExecuteExtScriptInFriendlyFrame();
+      assert.equal(output, '');
+    });
+
+    it('should have URL in returned value if url is passed', function () {
+      var url = 'https://abcd.com/service?a=1&b=2&c=3';
+      var output = utils.createContentToExecuteExtScriptInFriendlyFrame(url);
+      var expected = `<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd"><html><head><base target="_top" /><script>inDapIF=true;</script></head><body><!--PRE_SCRIPT_TAG_MACRO--><script src="${url}"></script><!--POST_SCRIPT_TAG_MACRO--></body></html>`;
+      assert.equal(output, expected);
+    });
+  });
+
+  describe('getDefinedParams', () => {
+    it('builds an object consisting of defined params', () => {
+      const adUnit = {
+        mediaType: 'video',
+        comeWithMe: 'ifuwant2live',
+        notNeeded: 'do not include',
+      };
+
+      const builtObject = utils.getDefinedParams(adUnit, [
+        'mediaType', 'comeWithMe'
+      ]);
+
+      assert.deepEqual(builtObject, {
+        mediaType: 'video',
+        comeWithMe: 'ifuwant2live',
+      });
     });
   });
 
