@@ -1,47 +1,59 @@
 var prebid = require('./package.json');
 var StringReplacePlugin = require('string-replace-webpack-plugin');
 var path = require('path');
+var webpack = require('webpack');
+var helpers = require('./gulpHelpers');
+var RequireEnsureWithoutJsonp = require('./plugins/RequireEnsureWithoutJsonp.js');
+
+// list of module names to never include in the common bundle chunk
+var neverBundle = [
+  'AnalyticsAdapter.js'
+];
 
 module.exports = {
-  output: {
-    filename: 'prebid.js'
-  },
   devtool: 'source-map',
   resolve: {
-    modulesDirectories: ['', 'node_modules', 'src']
+    modules: [
+      path.resolve('.'),
+      'node_modules'
+    ],
   },
-  resolveLoader: {
-    modulesDirectories: ['loaders', 'node_modules']
+  output: {
+    jsonpFunction: prebid.globalVarName+"Chunk"
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.js$/,
-        include: /(src|test)/,
-        exclude: path.resolve(__dirname, 'node_modules'),
-        loader: 'babel', // 'babel-loader' is also a legal name to reference
-        query: {
-          presets: ['es2015']
-        }
+        exclude: path.resolve('./node_modules'), // required to prevent loader from choking non-Prebid.js node_modules
+        use: [
+          {
+            loader: 'babel-loader',
+            options: {
+              presets: ['es2015']
+            }
+          }
+        ]
+      },
+      { // This makes sure babel-loader is ran on our intended Prebid.js modules that happen to be in node_modules
+        test: /\.js$/,
+        include: helpers.getArgModules().map(module => new RegExp('node_modules/' + module + '/')),
+        use: [
+          {
+            loader: 'babel-loader',
+            options: {
+              presets: ['es2015']
+            }
+          }
+        ],
       },
       {
         test: /\.json$/,
-        loader: 'json'
+        loader: 'json-loader'
       },
       {
-        test: /adaptermanager.js/,
-        include: /(src)/,
-        loader: 'analyticsLoader'
-      },
-      {
-        test: /adaptermanager.js/,
-        include: /(src)/,
-        loader: 'adapterLoader'
-      },
-      {
-        test: /native.js/,
-        include: /(src)/,
-        loader: 'nativeLoader'
+        test: /\.md$/,
+        loader: 'ignore-loader'
       },
       {
         test: /constants.json$/,
@@ -57,23 +69,33 @@ module.exports = {
           ]
         })
       },
-        {
-          test: /\.js$/,
-          include: /(src|test|integrationExamples)/,
-          loader: StringReplacePlugin.replace({
-            replacements: [
-              {
-                pattern: /\$\$PREBID_GLOBAL\$\$/g,
-                replacement: function (match, p1, offset, string) {
-                    return prebid.globalVarName;
-                }
+      {
+        test: /\.js$/,
+        include: /(src|test|modules|integrationExamples)/,
+        loader: StringReplacePlugin.replace({
+          replacements: [
+            {
+              pattern: /\$\$PREBID_GLOBAL\$\$/g,
+              replacement: function (match, p1, offset, string) {
+                return prebid.globalVarName;
               }
-            ]
-          })
-        }
+            }
+          ]
+        })
+      }
     ]
   },
   plugins: [
-    new StringReplacePlugin()
+    new StringReplacePlugin(),
+    new RequireEnsureWithoutJsonp(),
+
+    // this plugin must be last so it can be easily removed for karma unit tests
+    new webpack.optimize.CommonsChunkPlugin({
+      name: 'prebid',
+      filename: 'prebid-core.js',
+      minChunks: function(module, count) {
+        return !(count < 2 || neverBundle.includes(path.basename(module.resource)))
+      }
+    })
   ]
 };
